- fixed a typo, moved tests to same file
