+++ src/test/java/htsjdk/tribble/TribbleIndexFeatureReaderTest.java
@@ -1,6 +1,8 @@
 package htsjdk.tribble;
 
 import htsjdk.HtsjdkTest;
+import htsjdk.samtools.util.Interval;
+import htsjdk.tribble.IntervalList.IntervalListCodec;
 import htsjdk.tribble.readers.LineIterator;
 import htsjdk.variant.variantcontext.VariantContext;
 import htsjdk.variant.vcf.VCFCodec;
@@ -8,6 +10,7 @@ import org.testng.Assert;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
+import java.io.File;
 import java.io.IOException;
 
 
@@ -29,7 +32,34 @@ public class TribbleIndexFeatureReaderTest extends HtsjdkTest {
         try (final TribbleIndexedFeatureReader<VariantContext, LineIterator> featureReader =
                      new TribbleIndexedFeatureReader<>(testPath, codec, false)) {
 
-            Assert.assertEquals(featureReader.iterator().stream().count(), 1);
+            Assert.assertEquals(featureReader.iterator().stream().count(), expectedCount);
+        }
+    }
+
+
+    @DataProvider()
+    public Object[][] createIntervalFileStrings() {
+        return new Object[][]{
+                {new File(TestUtils.DATA_DIR, "interval_list/shortExample.interval_list"), 4}
+        };
+    }
+
+    @Test(dataProvider = "createFeatureFileStrings")
+    public void testIndexedIntervalList(final File testPath, final int expectedCount) throws IOException {
+        final IntervalListCodec codec = new IntervalListCodec();
+        try (final TribbleIndexedFeatureReader<Interval, LineIterator> featureReader =
+                     new TribbleIndexedFeatureReader<>(testPath.getAbsolutePath(), codec, false)) {
+            Assert.assertEquals(featureReader.iterator().stream().count(), expectedCount);
+        }
+    }
+
+    @Test(dataProvider = "createFeatureFileStrings", expectedExceptions = TribbleException.class)
+    public void testIndexedIntervalListWithQuery(final File testPath, final int ignored) throws IOException {
+        final IntervalListCodec codec = new IntervalListCodec();
+        try (final TribbleIndexedFeatureReader<Interval, LineIterator> featureReader =
+                     new TribbleIndexedFeatureReader<>(testPath.getAbsolutePath(), codec, false)) {
+
+            Assert.assertEquals(featureReader.query("1", 17032814, 17032814).stream().count(), 1);
         }
     }
 }
